+++ core/src/connection/mod.rs
@@ -112,7 +112,7 @@ pub async fn authenticate(
         _ => CpuFamily::CPU_UNKNOWN,
     };
 
-    let os = match std::env::consts::OS {
+    let os = match crate::config::OS {
         "android" => Os::OS_ANDROID,
         "freebsd" | "netbsd" | "openbsd" => Os::OS_FREEBSD,
         "ios" => Os::OS_IPHONE,